modifications to keep original aspect ratio
works when image is smaller than max image(nothing done)
works when height is == to max, and width is less
works when height is == to max, and width is more
works when width is == to max, and height is less
works when width is == to max, and height is more
works when aspect ratio = 1 (square)
works when both width and height are larger than max

Aspect ratio is kept in all conditions.  This means, not ugly white
background when image is resized to within maximum dimensions setting.